requirements-working-bak.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.39.4
botocore==1.39.4
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
distro==1.9.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
oauthlib==3.3.1
openai==1.95.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
